LPS-55267 Moves code from EditGroupAction to portlet, no logic changes
